+++ babel.config.cjs
@@ -1,8 +1,6 @@
-module.exports = async () => {
-  return {
-    presets: [
-      ['@babel/preset-env', { targets: { node: 'current' } }],
-      '@babel/preset-typescript',
-    ],
-  };
+module.exports = {
+  presets: [
+    ['@babel/preset-env', { targets: { node: 'current' } }],
+    '@babel/preset-typescript',
+  ],
 };