+++ declaration.d.ts
@@ -1,9 +1,11 @@
 declare module 'compareq' {
-  export default function compare(a: any, b: any): boolean;
-}
-
-declare module 'nanoid-esm/non-secure' {
-  export default function nanoid(): string;
+  /**
+   * Deep comparison of two values
+   * @param a First value to compare
+   * @param b Second value to compare
+   * @returns True if values are equal
+   */
+  export default function compare<T>(a: T, b: T): boolean;
 }
 
 declare module 'rollup-plugin-uglify';