+++ src/index.ts
@@ -1,6 +1,7 @@
 import connector, { Config } from './connector';
 
-export { createStore } from './create';
+export { createStore, StoreShape, InferStoreType } from './create';
+export type { InferStoreFromConfig, Config } from './connector';
 
 /**
  * Creates a connection to a store or stores specified in the configuration
@@ -11,6 +12,7 @@ export { createStore } from './create';
  * @returns A disconnect function that can be used to unsubscribe the callback function from changes in the store(s).
  */
 export function connect(config: Config, cb: Function): Function {
+  // console.log('Entering connect function:', config, cb);
   const { connect, disconnect } = connector(config, cb);
 
   connect(); // immediately connect