+++ src/plugins/preact.ts
@@ -1,44 +1,16 @@
-import { h, Component } from 'preact';
+import { h } from 'preact';
+import { useState, useEffect, useCallback, useMemo } from 'preact/hooks';
 
 import connector from '../connector';
 import { createConnectedStore } from '../create';
 
 import { createStore as _createStore } from '../';
 
-function createUpdater(instance) {
+function createUpdater(instance: any) {
   return () => instance.setState({ _justorm: Date.now() });
 }
 
-export function createStore(instance, obj) {
+export function createStore(instance: any, obj: any) {
   if (typeof instance === 'string') return _createStore(instance, obj);
   return createConnectedStore(obj, createUpdater(instance));
-}
-
-export function withStore(config = {}) {
-  return function (WrappedComponent) {
-    return class extends Component {
-      constructor(props) {
-        super(props);
-
-        this.state = { updated: null };
-        this.update = createUpdater(this);
-        this.justormAPI = connector(config, this.update);
-
-        this.justormAPI.connect();
-      }
-
-      componentWillUnmount() {
-        this.justormAPI.disconnect();
-      }
-
-      toString() {
-        return `withStore(${WrappedComponent.name})`;
-      }
-
-      render(props) {
-        const { store } = this.justormAPI;
-        return h(WrappedComponent, Object.assign(props, { store }));
-      }
-    };
-  };
 }