+++ tests/index.test.ts
@@ -1,19 +1,8 @@
-import { spyOn } from 'jest-mock';
-import { describe, expect, it } from '@jest/globals';
+import { describe, expect, it, jest } from '@jest/globals';
 
-import { createStore, connect } from '../dist/esm/src';
+import { createStore, connect } from '../src';
 
 describe('justorm', () => {
-  let consoleLogSpy;
-
-  beforeEach(() => {
-    consoleLogSpy = spyOn(console, 'log').mockImplementation();
-  });
-
-  afterEach(() => {
-    consoleLogSpy.mockRestore();
-  });
-
   describe('createStore', () => {
     it('initializes a store with a number field correctly', () => {
       const store = createStore('store', { a: 1 });
@@ -23,11 +12,15 @@ describe('justorm', () => {
     it('initializes a store with a string field correctly', () => {
       const store = createStore('store', { a: 'test' });
       expect(store.a).toEqual('test');
+
+      store.a = 'test';
     });
 
     it('initializes a store with an array field correctly', () => {
       const store = createStore('store', { a: [1, 2, 3] });
       expect(store.a).toEqual([1, 2, 3]);
+      store.a.push(4);
+      expect(store.a).toEqual([1, 2, 3, 4]);
     });
 
     it('initializes a store with an object field correctly', () => {
@@ -39,12 +32,14 @@ describe('justorm', () => {
       const map = new Map();
       const store = createStore('store', { a: map });
 
+      // @ts-ignore
       expect(store.a.originalObject).toEqual(map);
     });
 
     it('initializes a store with a Set field correctly', () => {
       const set = new Set();
       const store = createStore('store', { a: set });
+      // @ts-ignore
       expect(store.a.originalObject).toEqual(set);
     });
   });
@@ -52,13 +47,13 @@ describe('justorm', () => {
   describe('connect', () => {
     it('calls the callback when a number field changes', done => {
       const store = createStore('store', { a: 1 });
-      connect({ store: 'a' }, () => done());
+      connect({ store: ['a'] }, () => done());
       store.a = 2;
     });
 
     it('calls the callback when a string field changes', done => {
       const store = createStore('store', { a: 'test' });
-      connect({ store: 'a' }, () => done());
+      connect({ store: ['a'] }, () => done());
       store.a = 'changed';
     });
 
@@ -68,14 +63,16 @@ describe('justorm', () => {
   describe('disconnect', () => {
     it('prevents the callback from being called when a field changes', done => {
       const store = createStore('store', { a: 1 });
-      let callbackCalled = false;
-      const disconnect = connect({ store: 'a' }, () => (callbackCalled = true));
+      const callback = jest.fn();
+      const disconnect = connect({ store: ['a'] }, callback);
+
       disconnect();
       store.a = 2;
+
       setTimeout(() => {
-        expect(callbackCalled).toBe(false);
+        expect(callback).not.toHaveBeenCalled();
         done();
-      }, 50);
+      }, 100);
     });
   });
 });